backend/requirements.txt

dj-database-url==0.5.0
Django==3.1.5
django-allauth==0.44.0
django-cors-headers==3.7.0
django-filter==2.4.0
django-heroku==0.3.1
django-rest-auth==0.9.5
django-rest-swagger==2.2.0
djangorestframework==3.12.2
drf-yasg==1.20.0
asgiref==3.2.10
defusedxml==0.6.0
diff-match-patch==20181111
dj-database-url==0.5.0
et-xmlfile==1.0.1
gunicorn==20.0.4
jdcal==1.4.1
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.4
psycopg2==2.8.5
pytz==2020.1
PyYAML==5.3.1
sqlparse==0.3.1
tablib==2.0.0
whitenoise==5.1.0
xlrd==1.2.0
xlwt==1.3.0